FMbw {dipSFP}	R Documentation

Modified critical bandwidth

Description

Modified critical bandwidth in Fisher and Marron (2001).

Usage


FMbw(vec, k, lambda0 = NULL, outliers.ratio = NULL, m0 = NULL, tol = 1e-06)

Arguments

vec 	 a vector.

k 	 the number of modes.

lambda0,outliers.ratio,m0 	 tuning parameters, see `Details' below.

tol 	 the desired accuracy (convergence tolerance) in computing the modified critical bandwidth by bisection method.

Details

Fisher and Marron (2001) proposed a modified critical bandwidth instead of Silverman critical bandwidth (1981) to estimate the underlying data distribution and generate bootstrap resamples in smooth bootstrap algorithm. Two tuning parameters `lambda0' and `m0' were used to deal with the outliers problems and distinguish minor or major modes. As suggested in Fisher and Marron (2001), `lambda0' can be given directly by user or determined by the ratio of outliers `outliers.ratio'. If both of them are given, then the maximum of `lambda0' and `outliers.ratio/h/(2*pi)^(1/2)' is used in computation, where `h' is the bandwidth at which the statistic S_k is calculated. The defaults of `lambda0' and `outliers.ratio' are both zero. The default of `m_0' is 0.003.

Value

The value of modified critical bandwidth.

Author(s)

Na You

References

Fisher, N. I. and Marron, J. S. (2001). Mode testing via the excess mass estimate. Biometrika, 88(2):499–517.

Examples


x <- rnorm(100,0,1)
FMbw(x,1)
FMbw(x,1,m0=0)

[Package dipSFP version 1.0 Index]