dipSFP {dipSFP}	R Documentation

SFP detection procedure based on dip statistic

Description

SFP detection procedure based on dip statistic, where the null distribution of dip statistic is estimated by smooth bootstrap algorithm.

Usage


dipSFP(dat, nboot, lambda0 = NULL, outliers.ratio = NULL, m0 = NULL,
pvalue = TRUE, tol = 1e-06)

Arguments

dat 	 a matrix of microarray data to be analyzed, for example, the estimated binding affynity values or SFPdev values, where each row corresponds to one probe and each column corresponds to one chip.

nboot 	 the number of bootstrap replications.

lambda0,outliers.ratio,m0 	 tuning parameters, see `Details' in FMbw.

pvalue 	 logical; will p values be computed?

tol 	 the desired accuracy (convergence tolerance) in computing the modified critical bandwidth by bisection method.

Details

Dip statistic is computed by dip in R package `diptest'.

Value

A list with four components if pvalue is TRUE,

dip.statistics 	 dip statistic values for probes

null.distribution 	 estimated null distribution of dip statistic

p.values 	 p values

adjusted.pvalues 	 adjusted p values using the false discovery rate criterion

If pvalue is FALSE, only dip.statistics and null.distribution are returned.

Author(s)

Na You

References

Cui et al. Single Feature Polymorhpism Detection using Recombinant Inbred Line microarray expression data.

Fisher, N. I. and Marron, J. S. (2001). Mode testing via the excess mass estimate. Biometrika, 88(2):499–517.

Hartigan, J. A. and Hartigan, P. M. (1985). The dip test of unimodality. The Annals of Statistics, 13(1):70–84.

See Also

FMbw

Examples


n <- 100
m.nonSFP <- 98
m.SFP <- 2

dat1 <- matrix(rgamma(m.nonSFP*n,2,4),m.nonSFP,n)
dat2 <- cbind(matrix(rgamma(m.SFP*n/2,2,4),m.SFP,n/2),
              matrix(rgamma(m.SFP*n/2,20,4),m.SFP,n/2))
dat <- rbind(dat1,dat2)

dipSFP(dat,nboot=20)
dipSFP(dat,nboot=20,outliers.ratio=0.01)

[Package dipSFP version 1.0 Index]