dipSFP-package {dipSFP}	R Documentation

SFP detection based on dip statistic

Description

Given a matrix of microarray data, SFPs are detected based on the dip statistic. The null distribution of dip statistic is estimated by smooth bootstrap algorithm. P values are adjusted using false discovery rate criteria for FDR control.

Details

Package: 	 dipSFP

Type: 	 Package

Version: 	 1.0

Date: 	 2010-05-05

LazyLoad: 	 yes

Author(s)

Na You

Maintainer: Na You <nayou@ucr.edu>

References

Cui et al. Single Feature Polymorhpism Detection using Recombinant Inbred Line microarray expression data.

[Package dipSFP version 1.0 Index]