minimax {dipSFP}	R Documentation

Minimax function

Description

Determine if each point in a line is a local maximum or minimum.

Usage


minimax(func)

Arguments

func 	 A matrix. Each row indicates the (x,y)-coordinate of one point.

Value

A vector with the same length as the first dimension of matrix func, composed of -1, 1, and 0, indicating respectively the corresponding point is a local minimum, maximum or none of them.

Author(s)

Na You

References

Fisher, N. I. and Marron, J. S. (2001). Mode testing via the excess mass estimate. Biometrika, 88(2):499–517.

Examples


x <- rnorm(10,0,1)
minimax(cbind(seq(x),x))

[Package dipSFP version 1.0 Index]