modes {dipSFP}	R Documentation

Modes of density function

Description

Given a vector of data, find the `major' modes of its Gaussian kernal density estimate with specified bandwidth, exluding the modes caused by outliers and `minor' modes with small masses.

Usage


modes(vec, bw, lambda0 = NULL, outliers.ratio = NULL, m0 = NULL)

Arguments

vec 	 a vector.

bw 	 bandwidth used in the Gaussion kernal density estimate.

lambda0, outliers.ratio, m0 	 tuning parameters. See `Detailes' in FMbw.

Value

A vector, each element of which indicates the location of one mode.

Author(s)

Na You

References

Fisher, N. I. and Marron, J. S. (2001). Mode testing via the excess mass estimate. Biometrika, 88(2):499–517.

See Also

FMbw

Examples


x <- rnorm(100,0,1)
h1 <- FMbw(x,k=1)
modes(x,bw=h1)

h2 <- FMbw(x,k=2)
modes(x,bw=h2)

[Package dipSFP version 1.0 Index]